Change badtolerance to long

+++ community/import-tool/src/main/java/org/neo4j/tooling/ImportTool.java
@@ -373,7 +373,7 @@ public class ImportTool
         boolean enableStacktrace;
         Number processors = null;
         Input input = null;
-        int badTolerance;
+        long badTolerance;
         Charset inputEncoding;
         boolean skipBadRelationships, skipDuplicateNodes, ignoreExtraColumns;
         boolean skipBadEntriesLogging;
@@ -512,7 +512,7 @@ public class ImportTool
         finally
         {
             Collector collector = input.badCollector();
-            int numberOfBadEntries = collector.badEntries();
+            long numberOfBadEntries = collector.badEntries();
             collector.close();
             IOUtils.closeAll( badOutput );
 
@@ -572,7 +572,7 @@ public class ImportTool
         };
     }
 
-    private static Collector getBadCollector( int badTolerance, boolean skipBadRelationships,
+    private static Collector getBadCollector( long badTolerance, boolean skipBadRelationships,
             boolean skipDuplicateNodes, boolean ignoreExtraColumns, boolean skipBadEntriesLogging,
             OutputStream badOutput )
     {
@@ -580,10 +580,10 @@ public class ImportTool
         return skipBadEntriesLogging ? silentBadCollector( badTolerance, collect ) : badCollector( badOutput, badTolerance, collect );
     }
 
-    private static Integer parseNumberOrUnlimited( Args args, Options option )
+    private static long parseNumberOrUnlimited( Args args, Options option )
     {
         String value = args.get( option.key(), option.defaultValue().toString() );
-        return UNLIMITED.equals( value ) ? BadCollector.UNLIMITED_TOLERANCE : Integer.parseInt( value );
+        return UNLIMITED.equals( value ) ? BadCollector.UNLIMITED_TOLERANCE : Long.parseLong( value );
     }
 
     private static Config loadDbConfig( File file ) throws IOException
