+++ community/kernel/src/test/java/org/neo4j/unsafe/impl/batchimport/cache/idmapping/string/EncodingIdMapperTest.java
@@ -842,7 +842,7 @@ public class EncodingIdMapperTest
         }
 
         @Override
-        public int badEntries()
+        public long badEntries()
         {
             throw new UnsupportedOperationException();
         }
